ver 0.12 Add Stopwatches

+++ MultithreadedSimulation_lab2/Model/PhilosopherMetrics.cs
@@ -3,7 +3,7 @@
 public class PhilosopherMetrics
 {
     public int Eaten { get; private set; }
-    public int WaitingTime { get; set; }
+    public long WaitingTime { get; set; }
     
     public void IncrementEaten() {Eaten++;}
 }
